GH-128 Implement flags for user

+++ Game-Hub/game-Hub-be/game-Hub/src/main/java/gamehub/game_Hub/Repository/UserStoreFlagRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import gamehub.game_Hub.Module.Flags.UserStoreFlag;
 
-public interface UserStoreFlagRepository extends JpaRepository<UserStoreFlag, Integer> {
+public interface UserStoreFlagRepository extends JpaRepository<UserStoreFlag, Long> {
 
 }
